Move maybe_trim_page_path and maybe_add_page_path functions to context_util
