+++ lib/tests/streamlit/runtime/context_test.py
@@ -22,11 +22,7 @@ from parameterized import parameterized
 from tornado.httputil import HTTPHeaders
 
 import streamlit as st
-from streamlit.runtime.context import (
-    _normalize_header,
-    maybe_add_page_path,
-    maybe_trim_page_path,
-)
+from streamlit.runtime.context import _normalize_header
 
 
 class StContextTest(unittest.TestCase):
@@ -102,21 +98,21 @@ class StContextTest(unittest.TestCase):
         mock_ctx.pages_manager = MagicMock()
 
         # Set up the mock return values for the URL processing functions
-        mock_trim_path.return_value = "https://example.com/trimmed"
-        mock_add_path.return_value = "https://example.com/trimmed/added"
+        mock_trim_path.return_value = "https://example.com/"
+        mock_add_path.return_value = "https://example.com/added"
 
         # Test that the URL is processed by both functions
         result = st.context.url
 
         # Verify the result
-        assert result == "https://example.com/trimmed/added"
+        assert result == "https://example.com/added"
 
         # Verify that the functions were called with the correct arguments
         mock_trim_path.assert_called_once_with(
             "https://example.com/original", mock_ctx.pages_manager
         )
         mock_add_path.assert_called_once_with(
-            "https://example.com/trimmed", mock_ctx.pages_manager
+            "https://example.com/", mock_ctx.pages_manager
         )
 
     @parameterized.expand(
@@ -137,111 +133,3 @@ class StContextTest(unittest.TestCase):
     def test_normalize_header(self, name, expected):
         """Test that `_normalize_header` normalizes header names"""
         assert _normalize_header(name) == expected
-
-    @parameterized.expand(
-        [
-            # Test case: URL with no page path
-            ("https://example.com", {}, "https://example.com"),
-            # Test case: URL with page path that matches a page
-            (
-                "https://example.com/page1",
-                {"hash1": {"url_pathname": "page1"}},
-                "https://example.com/",
-            ),
-            # Test case: URL with page path that doesn't match any page
-            (
-                "https://example.com/unknown",
-                {"hash1": {"url_pathname": "page1"}},
-                "https://example.com/unknown",
-            ),
-            # Test case: URL with trailing slash
-            (
-                "https://example.com/page1/",
-                {"hash1": {"url_pathname": "page1"}},
-                "https://example.com/",
-            ),
-            # Test case: URL with multiple segments where the last segment matches a page
-            (
-                "https://example.com/path/to/page1",
-                {"hash1": {"url_pathname": "page1"}},
-                "https://example.com/path/to/",
-            ),
-            # Test case: URL with empty page path in pages
-            (
-                "https://example.com",
-                {"hash1": {"url_pathname": ""}},
-                "https://example.com",
-            ),
-            # Test case: URL with multiple pages, one matching
-            (
-                "https://example.com/page2",
-                {
-                    "hash1": {"url_pathname": "page1"},
-                    "hash2": {"url_pathname": "page2"},
-                    "hash3": {"url_pathname": "page3"},
-                },
-                "https://example.com/",
-            ),
-        ]
-    )
-    def test_maybe_trim_page_path(self, url, pages, expected):
-        """Test that `maybe_trim_page_path` correctly trims page paths from URLs"""
-        # Create a mock PagesManager
-        mock_page_manager = MagicMock()
-        mock_page_manager.get_pages.return_value = pages
-
-        # Call the function and check the result
-        result = maybe_trim_page_path(url, mock_page_manager)
-        assert result == expected
-
-    @parameterized.expand(
-        [
-            # Test case: URL with no current page
-            ("https://example.com", "", {}, "https://example.com"),
-            # Test case: URL with the current page that has a url_pathname
-            (
-                "https://example.com",
-                "hash1",
-                {"hash1": {"url_pathname": "page1"}},
-                "https://example.com/page1",
-            ),
-            # Test case: URL with current page that has no url_pathname
-            (
-                "https://example.com",
-                "hash1",
-                {"hash1": {"page_name": "Page 1"}},
-                "https://example.com",
-            ),
-            # Test case: URL with current page that has empty url_pathname
-            (
-                "https://example.com",
-                "hash1",
-                {"hash1": {"url_pathname": ""}},
-                "https://example.com",
-            ),
-            # Test case: URL with trailing slash
-            (
-                "https://example.com/",
-                "hash1",
-                {"hash1": {"url_pathname": "page1"}},
-                "https://example.com/page1",
-            ),
-            # Test case: URL with the current page hash that doesn't exist in pages
-            (
-                "https://example.com",
-                "unknown",
-                {"hash1": {"url_pathname": "page1"}},
-                "https://example.com",
-            ),
-        ]
-    )
-    def test_maybe_add_page_path(self, url, current_hash, pages, expected):
-        """Test that `maybe_add_page_path` correctly adds page paths to URLs"""
-        # Create a mock PagesManager
-        mock_page_manager = MagicMock()
-        mock_page_manager.current_page_script_hash = current_hash
-        mock_page_manager.get_pages.return_value = pages
-
-        # Call the function and check the result
-        result = maybe_add_page_path(url, mock_page_manager)
-        assert result == expected
